creating api and attribute factory

+++ src/main/java/domain/TopClothing.java
@@ -9,7 +9,7 @@ public class TopClothing extends Clothing {
 
     }
 
-    public TopClothing(int userID, String name) {
+    public TopClothing(long userID, String name) {
         super(userID, name);
     }
 }
